# Use ngx_posted_delayed_events implement 0-delay timers