--- a/src/ngx_http_lua_timer.c
+++ b/src/ngx_http_lua_timer.c
@@ -351,10 +351,20 @@ ngx_http_lua_ngx_timer_helper(lua_State *L, int every)
 
     lmcf->pending_timers++;
 
-    ngx_add_timer(ev, delay);
+    if (delay == 0) {
+#ifdef HAVE_POSTED_DELAYED_EVENTS_PATCH
+        dd("posting 0 sec sleep event to head of delayed queue");
 
-    ngx_log_debug2(NGX_LOG_DEBUG_HTTP, ngx_cycle->log, 0,
+        ngx_post_event(ev, &ngx_posted_delayed_events);
+#else
+        ngx_add_timer(ev, delay);
+#endif
+
+    } else {
+        ngx_add_timer(ev, delay);
+        ngx_log_debug2(NGX_LOG_DEBUG_HTTP, ngx_cycle->log, 0,
                    "created timer (co: %p delay: %M ms)", tctx->co, delay);
+    }
 
     lua_pushinteger(L, 1);
     return 1;